Fix RNG interfacing with JS

+++ src/lib.rs
@@ -18,9 +18,9 @@ pub struct WASMWorld {
 #[wasm_bindgen]
 impl WASMWorld {
     #[wasm_bindgen(constructor)]
-    pub fn new(volume_size: f64, dt: f64, seed: u64) -> Self {
+    pub fn new(volume_size: f64, dt: f64, seed: u32) -> Self {
         let volume = Volume::new(volume_size);
-        WASMWorld { world: World::new(vec![], volume, dt, seed) }
+        WASMWorld { world: World::new(vec![], volume, dt, seed as u64) }
     }
 
     pub fn step(&mut self) {
